edna-data/requirements.txt
pandas>=1.3.0
biopython>=1.79
argparse>=1.4.0